requirements.txt
ccxt>=4.1.13
ccxt-pro>=4.1.13
aiohttp>=3.8.5
pyyaml>=6.0.1
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
loguru>=0.7.2
python-telegram-bot==20.7
requests==2.31.0 